+++ packages/next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shallot-ui/next
 
+## 4.0.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/platform-react@2.0.0-beta.14
+  - @shallot-ui/platform-nextjs@2.0.0-beta.14
+
 ## 4.0.0-beta.14
 
 ### Patch Changes