+++ platforms/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/platform-react
 
+## 2.0.0-beta.14
+
+### Patch Changes
+
+- Improved merging of theme globals and variants
+
 ## 2.0.0-beta.13
 
 ### Patch Changes